refactor spikes sub-module
