+++ bycycle/__init__.py
@@ -1,4 +1,5 @@
 """ByCycle: cycle-by-cycle analyses of neural oscillations."""
 
 from .version import __version__
-from .objs import Bycycle, BycycleGroup, Spike
+from .objs import Bycycle, BycycleGroup
+from .spikes.objs import Spike
